moved enum to fix test failure
